requirements.txt
fastapi~=0.104
httpx~=0.25.1
psycopg[binary]~=3.1
python-dotenv~=1.0
pytest~=7.4
pytest-mock~=3.12
PyYAML~=6.0
SQLAlchemy~=2.0
uvicorn~=0.23
